Switch color-coded marker images from legacy API to new SVG code
This addresses a number of concerns and purposes.

This moves away from the already long-deprecated Google Image Charts
API, which was even once discontinued before being brought back a few
days later.

SVG XML code is easier to inline and generate within the javascript
front end, thus eliminating the need for a back end API to store or
generate images for all possible colors, and also for serving them
across the network: this reduces the requirements and strain on the
infrastructure, and improves performance.

Last but not least, SVG naturally provides support for high-DPI setups.