Discriminate between receivers according to blocked spectrum metric
Some KiwiSDR receivers, although technically capable of offering
access to the full 0-30 MHz range, are configured to block for example
everything but known amateur or broadcast bands. A receiver featuring
this kind of extensive blocking cannot be accurately considered as
wideband; milder blocking merely adjusts the precedence of the receiver
in proportion.

As with the earlier signal-to-noise ratio metric, the data source for
these spectrum metrics is still left out for now.